create abstraction
