+++ packages/grafana-ui/src/components/DataLinks/codemirrorUtils.ts
@@ -1,153 +1,46 @@
-import { Completion, CompletionContext, CompletionResult } from '@codemirror/autocomplete';
+import { autocompletion, Completion, CompletionContext, CompletionResult } from '@codemirror/autocomplete';
 import { Extension } from '@codemirror/state';
-import { Decoration, DecorationSet, EditorView, ViewPlugin, ViewUpdate } from '@codemirror/view';
+import { EditorView } from '@codemirror/view';
 
 import { DataLinkBuiltInVars, GrafanaTheme2, VariableOrigin, VariableSuggestion } from '@grafana/data';
 
+import { createGenericHighlighter } from '../CodeMirror/highlight';
+import { createGenericTheme } from '../CodeMirror/styles';
+
 /**
- * Creates a CodeMirror theme for data link input based on Grafana's theme
+ * Creates a CodeMirror theme for data link input with custom variable styling
+ * This extends the generic theme with data link-specific styles
  */
 export function createDataLinkTheme(theme: GrafanaTheme2): Extension {
-  const isDark = theme.colors.mode === 'dark';
+  const genericTheme = createGenericTheme(theme);
 
-  return EditorView.theme(
-    {
-      '&': {
-        fontSize: theme.typography.body.fontSize,
-        fontFamily: theme.typography.fontFamilyMonospace,
-        backgroundColor: 'transparent',
-        border: 'none',
-        outline: 'none',
-      },
-      '.cm-placeholder': {
-        color: theme.colors.text.disabled,
-        fontStyle: 'normal',
-      },
-      '.cm-scroller': {
-        overflow: 'auto',
-        fontFamily: theme.typography.fontFamilyMonospace,
-      },
-      '.cm-content': {
-        padding: '3px 0',
-        color: theme.colors.text.primary,
-        caretColor: theme.colors.text.primary,
-      },
-      '.cm-line': {
-        padding: '0 2px',
-      },
-      '.cm-cursor': {
-        borderLeftColor: theme.colors.text.primary,
-      },
-      '.cm-selectionBackground': {
-        backgroundColor: `${theme.colors.action.selected} !important`,
-      },
-      '&.cm-focused .cm-selectionBackground': {
-        backgroundColor: `${theme.colors.action.focus} !important`,
-      },
-      '.cm-variable': {
-        color: theme.colors.success.text,
-        fontWeight: theme.typography.fontWeightMedium,
-      },
-      '.cm-activeLine': {
-        backgroundColor: 'transparent',
-      },
-      '.cm-gutters': {
-        display: 'none',
-      },
-      '.cm-tooltip': {
-        zIndex: theme.zIndex.portal + 1, // Above modals and portals (1062)
-      },
-      '.cm-tooltip.cm-tooltip-autocomplete': {
-        backgroundColor: theme.colors.background.primary,
-        border: `1px solid ${theme.colors.border.weak}`,
-        boxShadow: theme.shadows.z3,
-      },
-      '.cm-tooltip.cm-tooltip-autocomplete > ul': {
-        fontFamily: theme.typography.fontFamily,
-        maxHeight: '300px',
-      },
-      '.cm-tooltip.cm-tooltip-autocomplete > ul > li': {
-        padding: '2px 8px',
-        color: theme.colors.text.primary,
-      },
-      '.cm-tooltip-autocomplete ul li[aria-selected]': {
-        backgroundColor: theme.colors.background.secondary,
-        color: theme.colors.text.primary,
-      },
-      '.cm-completionLabel': {
-        fontFamily: theme.typography.fontFamilyMonospace,
-        fontSize: theme.typography.size.sm,
-      },
-      '.cm-completionDetail': {
-        color: theme.colors.text.secondary,
-        fontStyle: 'normal',
-        marginLeft: theme.spacing(1),
-      },
-      '.cm-completionInfo': {
-        backgroundColor: theme.colors.background.primary,
-        border: `1px solid ${theme.colors.border.weak}`,
-        color: theme.colors.text.primary,
-        padding: theme.spacing(1),
-      },
+  // Add data link-specific variable styling
+  const dataLinkStyles = EditorView.theme({
+    '.cm-variable': {
+      color: theme.colors.success.text,
+      fontWeight: theme.typography.fontWeightMedium,
     },
-    { dark: isDark }
-  );
+  });
+
+  return [genericTheme, dataLinkStyles];
 }
 
 /**
  * Creates a syntax highlighter for data link variables (${...})
  * Matches the pattern from the old Prism implementation: (\${\S+?})
  */
-export function createDataLinkHighlighter(theme: GrafanaTheme2): Extension {
+export function createDataLinkHighlighter(): Extension {
   // Regular expression matching ${...} patterns (same as old implementation)
   const variablePattern = /\$\{[^}]+\}/g;
 
-  const variableDecoration = Decoration.mark({
-    class: 'cm-variable',
+  return createGenericHighlighter({
+    pattern: variablePattern,
+    className: 'cm-variable',
   });
-
-  const viewPlugin = ViewPlugin.fromClass(
-    class {
-      decorations: DecorationSet;
-
-      constructor(view: EditorView) {
-        this.decorations = this.buildDecorations(view);
-      }
-
-      update(update: ViewUpdate) {
-        if (update.docChanged || update.viewportChanged) {
-          this.decorations = this.buildDecorations(update.view);
-        }
-      }
-
-      buildDecorations(view: EditorView): DecorationSet {
-        const decorations: Array<{ from: number; to: number }> = [];
-        const text = view.state.doc.toString();
-        let match;
-
-        // Reset regex state
-        variablePattern.lastIndex = 0;
-
-        while ((match = variablePattern.exec(text)) !== null) {
-          decorations.push({
-            from: match.index,
-            to: match.index + match[0].length,
-          });
-        }
-
-        return Decoration.set(decorations.map((range) => variableDecoration.range(range.from, range.to)));
-      }
-    },
-    {
-      decorations: (v) => v.decorations,
-    }
-  );
-
-  return viewPlugin;
 }
 
 /**
- * Creates autocomplete function for data link variables
+ * Creates autocomplete source function for data link variables
  * Triggers on $ and = characters
  */
 export function dataLinkAutocompletion(
@@ -238,4 +131,18 @@ export function dataLinkAutocompletion(
       options,
     };
   };
+}
+
+/**
+ * Creates a data link autocompletion extension with configured suggestions
+ */
+export function createDataLinkAutocompletion(suggestions: VariableSuggestion[]): Extension {
+  return autocompletion({
+    override: [dataLinkAutocompletion(suggestions)],
+    activateOnTyping: true,
+    closeOnBlur: true,
+    maxRenderedOptions: 100,
+    defaultKeymap: true,
+    interactionDelay: 0,
+  });
 }
